LayoutTests/inspector/debugger/file-system-project-mapping.html

<html>
<head>
<script src="../../http/tests/inspector/inspector-test.js"></script>
<script src="../../http/tests/inspector/debugger-test.js"></script>
<script src="../../http/tests/inspector/workspace-test.js"></script>
<script src="../../http/tests/inspector/isolated-filesystem-test.js"></script>
<script>
function test()
{
    function dumpUISourceCodes(uiSourceCodes, next)
    {
        innerDumpUISourceCodes(uiSourceCodes, 0, next);

        function innerDumpUISourceCodes(uiSourceCodes, startIndex, next)
        {
            InspectorTest.addResult("");
            if (startIndex === uiSourceCodes.length) {
                next();
                return;
            }
            InspectorTest.dumpUISourceCode(uiSourceCodes[startIndex], innerDumpUISourceCodes.bind(this, uiSourceCodes, startIndex + 1, next));
        }
    }

    var manager;
    var resourceScriptMapping;
    var defaultScriptMapping;
    function createObjects()
    {
        InspectorTest.createWorkspace();
        manager = InspectorTest.createIsolatedFileSystemManager(InspectorTest.testWorkspace, InspectorTest.testFileSystemMapping);
        resourceScriptMapping = new WebInspector.ResourceScriptMapping(InspectorTest.testWorkspace);
        defaultScriptMapping = new WebInspector.DefaultScriptMapping(InspectorTest.testWorkspace);
    }

    InspectorTest.runTestSuite([
        function testAutomaticMapping(next)
        {
            function uiSourceCodeAdded(uiSourceCode) { }

            function dumpFileSystemUISourceCodesMappings()
            {
                var uiSourceCodes = InspectorTest.testWorkspace.project(fileSystemProjectId).uiSourceCodes();
                InspectorTest.addResult("UISourceCode uri to url mappings:");
                for (var i = 0; i < uiSourceCodes.length; ++i)
                    InspectorTest.addResult("    " + uiSourceCodes[i].uri() + " -> " + uiSourceCodes[i].url);
            }

            var fileSystemPath = "/var/www";
            var fileSystemProjectId = WebInspector.FileSystemProjectDelegate.projectId(fileSystemPath);
            var files = {"/html/foo.js": "<foo content>", "/bar.js": "<bar content>"};

            createObjects();
            InspectorTest.testFileMapping.setMappingEntries([]);
            InspectorTest.waitForWorkspaceUISourceCodeAddedEvent(uiSourceCodeAdded, 4);
            InspectorTest.addResult("Adding file system.");
            manager.addMockFileSystem(fileSystemPath, files);
            InspectorTest.addResult("Adding network resource.");
            InspectorTest.addMockUISourceCodeToWorkspace("http://localhost/html/foo.js", WebInspector.resourceTypes.Script, "<foo content>");
            InspectorTest.addMockUISourceCodeToWorkspace("http://localhost/bar.js", WebInspector.resourceTypes.Script, "<foo content>");
            dumpFileSystemUISourceCodesMappings();
            var uiSourceCode = InspectorTest.testWorkspace.uiSourceCode(fileSystemProjectId, ["html", "foo.js"]);
            var networkUISourceCode = InspectorTest.testWorkspace.uiSourceCode("http://localhost", ["html", "foo.js"]);
            InspectorTest.override(WebInspector, "suggestReload", function() { });
            InspectorTest.addResult("Adding mapping between network and file system resources.");
            InspectorTest.testWorkspace.addMapping(networkUISourceCode, uiSourceCode, manager.fileSystemWorkspaceProvider);
            manager.removeMockFileSystem(fileSystemPath);

            InspectorTest.addResult("Emulate reloading inspector.");
            createObjects();
            InspectorTest.waitForWorkspaceUISourceCodeAddedEvent(uiSourceCodeAdded, 2);
            manager.addMockFileSystem(fileSystemPath, files);
            dumpFileSystemUISourceCodesMappings();
            InspectorTest.override(WebInspector, "suggestReload", function() { });
            InspectorTest.addResult("Removing mapping between network and file system resources.");
            uiSourceCode = InspectorTest.testWorkspace.uiSourceCode(fileSystemProjectId, ["html", "foo.js"]);
            InspectorTest.testWorkspace.removeMapping(uiSourceCode);
            manager.removeMockFileSystem(fileSystemPath);

            InspectorTest.addResult("Emulate reloading inspector.");
            createObjects();
            InspectorTest.waitForWorkspaceUISourceCodeAddedEvent(uiSourceCodeAdded, 2);
            manager.addMockFileSystem(fileSystemPath, files);
            dumpFileSystemUISourceCodesMappings();
            manager.removeMockFileSystem("/var/www");

            InspectorTest.testFileMapping.setMappingEntries([]);
            next();
        }
    ]);
};
</script>
</head>
<body onload="runTest()">
<p>Tests file system project.</p>
</body>
</html>
